Create separate job get success functional and integration tests
